requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
python-dotenv
requests
